Move utility functions out of DS.Store, and into their own modules for reuse across ember-data
