+++ packages/ember-data/lib/system/utils/store.js
@@ -1,0 +1,28 @@
+export function serializerFor(container, type, defaultSerializer) {
+  return container.lookup('serializer:'+type) ||
+                 container.lookup('serializer:application') ||
+                 container.lookup('serializer:' + defaultSerializer) ||
+                 container.lookup('serializer:-default');
+}
+
+export function serializerForAdapter(adapter, type) {
+  var serializer = adapter.serializer;
+  var defaultSerializer = adapter.defaultSerializer;
+  var container = adapter.container;
+
+  if (container && serializer === undefined) {
+    serializer = serializerFor(container, type.typeKey, defaultSerializer);
+  }
+
+  if (serializer === null || serializer === undefined) {
+    serializer = {
+      extract: function(store, type, payload) { return payload; }
+    };
+  }
+
+  return serializer;
+}
+
+export function _adapterRun(store, fn) {
+  return store._backburner.run(fn);
+}
